// src/types/partials/DataRequirement.ts
/* eslint-disable no-unused-vars */
/* eslint-disable no-trailing-spaces */
// This file is auto-generated by generate_types so do not edit manually

import { TExtension } from '../partials/Extension';
import { TCanonical } from '../simpleTypes/Canonical';
import { TCodeableConcept } from '../partials/CodeableConcept';
import { TReference } from '../partials/Reference';
import { TDataRequirementCodeFilter } from '../partials/DataRequirementCodeFilter';
import { TDataRequirementDateFilter } from '../partials/DataRequirementDateFilter';
import { TInt } from '../simpleTypes/Int';
import { TDataRequirementSort } from '../partials/DataRequirementSort';

export type TDataRequirement = {
    id?: String;
    extension?: TExtension[];
    type: String;
    profile?: TCanonical[];
    subjectCodeableConcept?: TCodeableConcept;
    subjectReference?: TReference;
    mustSupport?: String[];
    codeFilter?: TDataRequirementCodeFilter[];
    dateFilter?: TDataRequirementDateFilter[];
    limit?: TInt;
    sort?: TDataRequirementSort[];
};